Fix employee performance issue with lots of employees
Using chrome devtools profiling I found that it took about 4 seconds
to run the process and almost all of it was in the four methods
called for each employee in 'calculateProductivity' like
'industry.getEmployeeCreMultiplier' along with the same method for Cha,
Int, and Eff, about 1/4 for each.   These in turn call
researchTree.getEmployeeCreMultiplier().  Each of these then processes
the whole research tree looking for multipliers and creating lots
of objects along the way.

This change memoizes `getMultiplierHelper(propName: string)`, and
adds a call to `Industry.process()` to reset it each time it is
called.